Enhanced global features interceptor
